+++ src/App.tsx
@@ -1,7 +1,5 @@
 import { Container } from './layouts/GlobalLayout/style';
 
-import { createContext, useState } from 'react';
-
 import Header from './components/common/Header';
 import Main from './components/common/Main';
 import Dropdown from './components/common/Dropdown';
@@ -19,23 +17,9 @@ import ToastPopup from './components/ToastPopup';
 import useFetchProducts from './hooks/useFetchProducts';
 
 import { CATEGORIES, PRICE_SORT } from './constants/filter';
-import { CartItem } from './types/cart';
 import { Category, Order } from './types/product';
 
-interface CartItemContextProps {
-  cartItems: CartItem[];
-  setCartItems: React.Dispatch<React.SetStateAction<CartItem[]>>;
-}
-
-const InitialState: CartItemContextProps = {
-  cartItems: [],
-  setCartItems: () => {},
-};
-
-export const CartItemContext = createContext(InitialState);
-
 function App() {
-  const [cartItems, setCartItems] = useState<CartItem[]>([]);
   const { error, category, filterByCategory, loading, products, sort, setSorting, observerRef } =
     useFetchProducts();
 
@@ -51,42 +35,40 @@ function App() {
   };
 
   return (
-    <CartItemContext.Provider value={{ cartItems, setCartItems }}>
-      <Container>
-        <Header>
-          <HomeButton onClick={() => {}} />
-          <CartButton count={cartItems.length} onClick={() => {}} />
-        </Header>
-        <ToastPopup
-          isOpen={Boolean(error)}
-          message="오류가 발생했습니다. 잠시 후 다시 시도해 주세요."
-        />
-        <Main>
-          <ProductsContainer>
-            <Title />
-            <FilterContainer>
-              <Dropdown
-                options={CATEGORIES}
-                selectedOption={selectedCategoryOption}
-                optionChange={handleCategoryChange}
-              />
-              <Dropdown
-                options={PRICE_SORT}
-                selectedOption={selectedSortOption}
-                optionChange={handlePriceSortChange}
-              />
-            </FilterContainer>
-            <ProductsContent>
-              {products.map((product) => (
-                <ProductItem key={product.id} {...product} />
-              ))}
-              {loading && <Loading />}
-              <div ref={observerRef} id="observer" style={{ height: '10px' }}></div>
-            </ProductsContent>
-          </ProductsContainer>
-        </Main>
-      </Container>
-    </CartItemContext.Provider>
+    <Container>
+      <Header>
+        <HomeButton onClick={() => {}} />
+        <CartButton onClick={() => {}} />
+      </Header>
+      <ToastPopup
+        isOpen={Boolean(error)}
+        message="오류가 발생했습니다. 잠시 후 다시 시도해 주세요."
+      />
+      <Main>
+        <ProductsContainer>
+          <Title />
+          <FilterContainer>
+            <Dropdown
+              options={CATEGORIES}
+              selectedOption={selectedCategoryOption}
+              optionChange={handleCategoryChange}
+            />
+            <Dropdown
+              options={PRICE_SORT}
+              selectedOption={selectedSortOption}
+              optionChange={handlePriceSortChange}
+            />
+          </FilterContainer>
+          <ProductsContent>
+            {products.map((product) => (
+              <ProductItem key={product.id} {...product} />
+            ))}
+            {loading && <Loading />}
+            <div ref={observerRef} id="observer" style={{ height: '10px' }}></div>
+          </ProductsContent>
+        </ProductsContainer>
+      </Main>
+    </Container>
   );
 }
