+++ src/components/CartButton/index.tsx
@@ -3,17 +3,19 @@ import * as S from './style';
 import { ButtonHTMLAttributes } from 'react';
 
 import { CART } from '../../assets/images';
+import useCart from '../../hooks/useCart';
 
 interface CartButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
   onClick: () => void;
-  count: number;
 }
 
-const CartButton = ({ onClick, count, ...props }: CartButtonProps) => {
+const CartButton = ({ onClick, ...props }: CartButtonProps) => {
+  const { cartItems } = useCart();
+
   return (
     <S.CartButton {...props} onClick={onClick}>
       <img src={CART} alt="장바구니" />
-      <S.NumberOfCartItems>{count}</S.NumberOfCartItems>
+      <S.NumberOfCartItems>{cartItems.length}</S.NumberOfCartItems>
     </S.CartButton>
   );
 };
