+++ src/context/CartItemProvider.tsx
@@ -1,0 +1,24 @@
+import { PropsWithChildren, createContext, useState } from 'react';
+import { CartItem } from '../types/cart';
+
+interface CartItemContextProps {
+  cartItems: CartItem[];
+  setCartItems: React.Dispatch<React.SetStateAction<CartItem[]>>;
+}
+
+const InitialState: CartItemContextProps = {
+  cartItems: [],
+  setCartItems: () => {},
+};
+
+export const CartItemsContext = createContext(InitialState);
+
+export const CartItemProvider = ({ children }: PropsWithChildren) => {
+  const [cartItems, setCartItems] = useState<CartItem[]>([]);
+
+  return (
+    <CartItemsContext.Provider value={{ cartItems, setCartItems }}>
+      {children}
+    </CartItemsContext.Provider>
+  );
+};
