+++ src/layouts/GlobalLayout/index.tsx
@@ -1,9 +1,14 @@
+import { CartItemProvider } from '../../context/CartItemProvider';
 import * as S from './style';
 
 import { PropsWithChildren } from 'react';
 
 const GlobalLayout = ({ children }: PropsWithChildren) => {
-  return <S.Container>{children}</S.Container>;
+  return (
+    <CartItemProvider>
+      <S.Container>{children}</S.Container>
+    </CartItemProvider>
+  );
 };
 
 export default GlobalLayout;
